chore(metrics): refactor name
